pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "faker>=37.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
